pyproject.toml
[project]
name = "ai-image-recognition-hub"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe==0.10.14",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
]

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"
